source_code/code_synthesis/taxrag/requirements.txt
pdfplumber
pandas
chromadb
python-dotenv
gradio
openai
z3-solver
